cast dimensions to int so that Word won't collapse.

+++ docx.py
@@ -416,8 +416,8 @@ def picture(relationshiplist, picname, picdescription='No Description', pixelwid
     # OpenXML measures on-screen objects in English Metric Units
     # 1cm = 36000 EMUs
     emuperpixel = 12667
-    width = str(pixelwidth * emuperpixel * scale)
-    height = str(pixelheight * emuperpixel * scale)
+    width = str(int(pixelwidth * emuperpixel * scale))
+    height = str(int(pixelheight * emuperpixel * scale))
 
     # Set relationship ID to the first available
     picid = '2'
